Stock_risk_analysis/requirements.txt
Flask
APScheduler
pandas
requests
beautifulsoup4
python-dotenv
lxml
